Move xdr persistence test to xdr tests
